beautybuddy_app/lib/extra_video_text/requirements.txt
flask
transformers
torch
openai-whisper
pydub